Support recyclable cache keys
ActiveSupport 5.2.0 introduces the concept of
[recyclable cache keys](https://github.com/rails/rails/pull/29092),
which prevents a bunch of unnecessary keys being created and then having
to be evicted. This should reduce the memory overhead of keeping a Redis
or Memcache server up in order to support a cache. However, this caused
issues for us because of an unneccessary override of `#write`. I can't
really see much of a difference between our implementation and the one
[baked into AS::Cache::Store](https://github.com/rails/rails/blob/master/activesupport/lib/active_support/cache.rb#L438-L448),
other than this `:race_condition_ttl` code which seems unnecessary since
`Cache::Store` supports it under the hood.

This should fix issues that people are having with rails 5.2.0.rc1 and
redis-activesupport.